Add host event handler
