+++ .github/workflows/cideploy.yml
@@ -1,24 +1,24 @@
-name: CIDeploy 
+name: CIDeploy
 
 on:
   push:
-    branches: [ main ]
+    branches: [main, v3.x]
 
 jobs:
   build:
     runs-on: ubuntu-latest
 
     steps:
-    - uses: actions/checkout@v2
+      - uses: actions/checkout@v2
 
-    - name: docker login
-      uses: docker/login-action@v2
-      with:
-        username: ${{secrets.DOCKER_USER}}
-        password: ${{secrets.DOCKER_TOKEN}}
+      - name: docker login
+        uses: docker/login-action@v2
+        with:
+          username: ${{secrets.DOCKER_USER}}
+          password: ${{secrets.DOCKER_TOKEN}}
 
-    - name: Build the Docker image
-      run: docker build --build-arg mixpanelToken=${{secrets.MIXPANEL_TOKEN}} -t reviewpad/action:${{github.sha}} .
-      
-    - name: Docker Push
-      run: docker push reviewpad/action:${{github.sha}}
+      - name: Build the Docker image
+        run: docker build --build-arg mixpanelToken=${{secrets.MIXPANEL_TOKEN}} -t reviewpad/action:${{github.sha}} .
+
+      - name: Docker Push
+        run: docker push reviewpad/action:${{github.sha}}
