+++ .gitignore
@@ -15,3 +15,9 @@ action
 # Dependency directories (remove the comment below to include it)
 # vendor/
 .task/*
+
+# IDEA
+.vscode
+
+# OS
+.DS_store
