+++ run.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# Copyright 2022 Explore.dev Unipessoal Lda. All Rights Reserved.
+# Copyright (C) 2022 Explore.dev Unipessoal Lda. All Rights Reserved.
 # Use of this source code is governed by a license that can be
 # found in the LICENSE file.
